add accessibility and focus order as per bluelines
