+++ express/code/scripts/color-shared/spectrum/utils/a11y.js
@@ -277,3 +277,50 @@ export function announceToScreenReader(message, priority = 'polite', options = {
     region.textContent = text ? text + suffix : '';
   }, delay);
 }
+
+// ── Roving Tabindex ──────────────────────────────────────────────────
+
+const ROVING_INDEX_ATTR = 'data-roving-index';
+
+/**
+ * Attach roving tabindex keyboard navigation to a container.
+ * Arrow keys (Left/Right/Up/Down) cycle through elements; Home/End jump
+ * to the first/last element. Only the active element has tabindex="0".
+ *
+ * @param {HTMLElement} container — the parent element to listen on
+ * @param {HTMLElement[]} elements — the focusable children
+ * @param {number} [initialFocusIndex=0] — which element starts with tabindex="0"
+ */
+export function attachRovingTabIndex(container, elements, initialFocusIndex = 0) {
+  if (!elements.length) return;
+  const focusIndex = Math.max(0, Math.min(initialFocusIndex, elements.length - 1));
+  elements.forEach((el, index) => {
+    el.setAttribute('tabindex', index === focusIndex ? '0' : '-1');
+    el.setAttribute(ROVING_INDEX_ATTR, index.toString());
+  });
+  container.addEventListener('keydown', (e) => {
+    const target = e.target.closest(`[${ROVING_INDEX_ATTR}]`);
+    if (!target || !elements.includes(target)) return;
+    const currentIndex = parseInt(target.getAttribute(ROVING_INDEX_ATTR), 10);
+    if (Number.isNaN(currentIndex) || currentIndex < 0 || currentIndex >= elements.length) return;
+    let targetIndex = -1;
+    if (e.key === 'ArrowRight' || e.key === 'ArrowDown') {
+      e.preventDefault();
+      targetIndex = (currentIndex + 1) % elements.length;
+    } else if (e.key === 'ArrowLeft' || e.key === 'ArrowUp') {
+      e.preventDefault();
+      targetIndex = (currentIndex - 1 + elements.length) % elements.length;
+    } else if (e.key === 'Home') {
+      e.preventDefault();
+      targetIndex = 0;
+    } else if (e.key === 'End') {
+      e.preventDefault();
+      targetIndex = elements.length - 1;
+    }
+    if (targetIndex !== -1 && targetIndex !== currentIndex) {
+      elements[currentIndex].setAttribute('tabindex', '-1');
+      elements[targetIndex].setAttribute('tabindex', '0');
+      elements[targetIndex].focus();
+    }
+  });
+}
